add rate function
