+++ src/main.rs
@@ -1,4 +1,10 @@
-fn main() {
+#[tokio::main]
+async fn main() {
     println!("{:?}", cli::working());
+    let res = cli::rate("https://github.com/ljharb/qs", env!("GITHUB_TOKEN"))
+        .await
+        .unwrap();
+    println!("{:?}", res);
+    println!("{:?}", res.responsive());
     println!("Hello, world!");
 }
